% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ibm}
\alias{ibm}
\title{Run a single replicate individual-based simulation}
\usage{
ibm(ngenerations)
}
\arguments{
\item{ngenerations}{Number of generations (integer)}
}
\description{
Run a single replicate individual-based simulation
}
